Add Mask + optimisation

+++ Ramjet_Core/Core/Rendering/Renderer/Font/Font.cpp
@@ -4,7 +4,7 @@ namespace Core {
 
 	namespace Rendering {
 
-		Font::Font(std::string name, std::string filename, int size)
+		Font::Font(std::string name, std::string filename, float size)
 			: m_Name(name), m_Filename(filename), m_Size(size), m_Scale(960 / (size), 540 / (size / 2))
 		{
 			if (size != FONT_STD_SIZE)
