+++ Ramjet_Core/Core/Rendering/Renderer/Font/Font.h
@@ -16,14 +16,14 @@ namespace Core {
 		private:
 			ftgl::texture_atlas_t* m_FTAtlas;
 			ftgl::texture_font_t* m_FTFont;
-			unsigned int m_Size;
+			float m_Size;
 
 			Maths::vec2 m_Scale;
 			std::string m_Name;
 			std::string m_Filename; 
 
 		public:
-			Font(std::string name, std::string filename, int size);
+			Font(std::string name, std::string filename, float size);
 
 			inline ftgl::texture_font_t* getFTFont() const { return m_FTFont; }
 
@@ -33,7 +33,7 @@ namespace Core {
 			inline const Maths::vec2 getScale() const { return m_Scale; }
 			inline const std::string getName() const { return m_Name; }
 			inline const std::string getFilename() const { return m_Filename; }
-			inline const int getSize() const { return m_Size; }
+			inline const float getSize() const { return m_Size; }
 		};
 
 	}
